.streamlit/files/requirements.txt
streamlit==1.18.1
numpy==1.24.1
pillow==9.4.0
requests==2.28.2
beautifulsoup4==4.11.2
joblib==1.2.0
scikit-learn==1.2.0
xgboost==1.7.3
tensorflow==2.11.0
